+++ packages/builder/example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:example/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:theta/theta.dart';
 
@@ -5,11 +6,10 @@ Future<void> main() async {
   /// Initialize Theta instance.
   /// You can get an anonymous key at https://app.buildwiththeta.com
   await Theta.initialize(
-    connectionMode: ConnectionMode.cached,
+    connectionMode: ConnectionMode.continuous,
 
     /// Example key
-    anonKey:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyZWYiOiI1NTE1MTU3Ni1jZGFkLTQ2ZGUtYWI3Mi01OGEzNDY1ZjM2NGMiLCJhbm9uX2tleSI6IjE0MzRkZDU5LTAwOWYtNGQ2Zi05OWE5LTQ4N2Y3Yzk1ODZhMCIsImlhdCI6MTY4OTk2MzQzNCwiZXhwIjoxNzIxNTIxMDM0LCJpc3MiOiJodHRwczovL2J1aWxkd2l0aHRoZXRhLmNvbSJ9.q32FmW5YAqzIncjRw0KXBwDr3J12qFumt5DrbEeyTKk',
+    anonKey: publicKey,
   );
 
   runApp(const MyApp());
@@ -59,6 +59,8 @@ class _MyAppState extends State<MyApp> {
           body: UIBox(
             'Counter View',
             controller: _controller,
+
+            branchName: 'Version 2.0',
 
             /// [placeholder] is the widget displayed while the page is loading.
             placeholder: const CircularProgressIndicator(),