+++ packages/builder/lib/src/data/repositories/component_repository_impl.dart
@@ -1,4 +1,5 @@
 import 'package:either_dart/either.dart';
+import 'package:light_logger/light_logger.dart';
 import 'package:theta/src/data/datasources/component_service.dart';
 import 'package:theta/src/data/datasources/local_component_service.dart';
 import 'package:theta/src/data/models/get_page_response.dart';
@@ -11,43 +12,64 @@ class ComponentRepositoryImpl implements ComponentRepository {
     this._preloadFile,
     this._componentService,
     this._localComponentService,
+    this._isCacheEnabled,
   );
 
   final PreloadFile _preloadFile;
   final ComponentService _componentService;
   final LocalComponentService _localComponentService;
+  final bool _isCacheEnabled;
+
+  Future<GetPageResponseEntity> _get(
+          String componentName, String? branchName) async =>
+      await _componentService.getComponent(componentName, branchName);
+
+  Future<Either<Exception, GetPageResponseEntity>> _getPreloaded(
+      String componentName, String? branchName) async {
+    try {
+      return Right(
+          await _localComponentService.getPreloadedComponent(componentName));
+    } catch (e) {
+      return Left(
+          Exception('Error preloading component, message: ${e.toString()}'));
+    }
+  }
+
+  Future<Either<Exception, GetPageResponseEntity>> _getCached(
+      String componentName, String? branchName) async {
+    final cachedComponent = await _localComponentService.getLocalComponent(
+        componentName, branchName);
+    if (cachedComponent != null) {
+      return Right(cachedComponent);
+    }
+    final res = await _get(componentName, branchName);
+    await _localComponentService.saveResponse(componentName, branchName, res);
+    return Right(res);
+  }
 
   @override
   Future<Either<Exception, GetPageResponseEntity>> getComponent(
     String componentName,
     bool preloadAllowed,
+    String? branchName,
   ) async {
     try {
       if (_preloadFile.enabled && preloadAllowed) {
-        try {
-          return Right(await _localComponentService
-              .getPreloadedComponent(componentName));
-        } catch (e) {
-          return Left(Exception(
-              'Error preloading component, message: ${e.toString()}'));
-        }
+        return _getPreloaded(componentName, branchName);
       }
-      final cachedComponent =
-          await _localComponentService.getLocalComponent(componentName);
-      if (cachedComponent != null) {
-        return Right(cachedComponent);
-      }
-      final res = await _componentService.getComponent(componentName);
-      await _localComponentService.saveResponse(componentName, res);
-      return Right(res);
-    } catch (e) {
-      try {
-        await _localComponentService.clearCache();
-        final res = await _componentService.getComponent(componentName);
+
+      if (!_isCacheEnabled) {
+        final res = await _get(componentName, branchName);
         return Right(res);
-      } catch (e) {
-        return Left(Exception(e.toString()));
       }
+
+      return _getCached(componentName, branchName);
+    } catch (e) {
+      Logger.printError('Error fetching component: $e');
+
+      await _localComponentService.clearCache();
+      final res = await _get(componentName, branchName);
+      return Right(res);
     }
   }