+++ packages/builder/lib/src/data/repositories/styles_repository_impl.dart
@@ -1,4 +1,5 @@
 import 'package:either_dart/either.dart';
+import 'package:light_logger/light_logger.dart';
 import 'package:theta/src/data/datasources/get_styles.dart.dart';
 import 'package:theta/src/data/datasources/local_styles_service.dart';
 import 'package:theta/src/data/models/get_styles_response.dart';
@@ -10,11 +11,13 @@ class StylesRepositoryImpl implements StylesRepository {
     this._preloadFile,
     this._stylesService,
     this._localStylesService,
+    this._isCacheEnabled,
   );
 
   final PreloadFile _preloadFile;
   final StylesService _stylesService;
   final LocalStylesService _localStylesService;
+  final bool _isCacheEnabled;
 
   @override
   Future<Either<Exception, GetStylesResponseEntity>> getStyles(
@@ -28,14 +31,20 @@ class StylesRepositoryImpl implements StylesRepository {
               Exception('Error preloading styles, message: ${e.toString()}'));
         }
       }
+      if (!_isCacheEnabled) {
+        final res = await _stylesService.fetch();
+        return Right(res);
+      }
+
       final cachedStyles = await _localStylesService.getLocalStyles();
       if (cachedStyles != null) {
         return Right(cachedStyles);
       }
       final res = await _stylesService.fetch();
-     await _localStylesService.saveResponse(res);
+      await _localStylesService.saveResponse(res);
       return Right(res);
     } catch (e) {
+      Logger.printError('Error fetching styles: $e');
       try {
         await _localStylesService.clearCache();
         final res = await _stylesService.fetch();