+++ packages/models/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.1
+
+- [x] Feat: added `toJson` to ConversionEvent.
+
 ## 0.3.0
 
 - [x] Feat: TreeState has now a `isPreloaded` property.